# web/doc/swagger.yaml
openapi: 3.0.0
info:
  title: 'Accencture API'
  contact:
    email: vlad.veresklya@yandex.ru
  version: 1.0.0
paths:
...
  /api/order:
    get:
      tags:
        - Заказы
      summary: 'Найти все заказы'
      parameters:
        -
          name: per-page
          in: query
          description: 'Кол-во записей на странице'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
  '/api/order/{id}':
    get:
      tags:
        - Заказы
      summary: 'Найти заказы по id'
      parameters:
        -
          name: id
          in: path
          description: 'ID заказа'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
    put:
      tags:
        - Заказы
      summary: 'Обновить заказ'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                number_vagon:
                  description: 'Кол-во вагонов'
                  type: integer
                weight:
                  description: 'Кол-во тонн'
                  type: double
                date:
                  description: Дата
                  type: date
                status:
                  description: Статус
                  type: integer
                defect_percent:
                  description: 'Процент брака'
                  type: integer
                defect_weight:
                  description: 'Вес брака'
                  type: integer
                vagons:
                  description: Вагоны
                  type: json
                id_warehouse:
                  description: 'Id склада'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Заказы
      summary: 'Удалить заказ по id'
      operationId: d1a919910bb61ace220d5b0f42230985
      parameters:
        -
          name: id
          in: path
          description: 'ID заказа'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
  /api/order/:
    post:
      tags:
        - Заказы
      summary: 'Добавить заказ'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                number_vagon:
                  description: 'Кол-во вагонов'
                  type: integer
                weight:
                  description: 'Кол-во тонн'
                  type: double
                date:
                  description: Дата
                  type: date
                status:
                  description: Статус
                  type: integer
                defect_percent:
                  description: 'Процент брака'
                  type: integer
                defect_weight:
                  description: 'Вес брака'
                  type: integer
                vagons:
                  description: Вагоны
                  type: json
                id_warehouse:
                  description: 'Id склада'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
  /api/reception-point:
    get:
      tags:
        - 'Пункты приема'
      summary: 'Найти все пункты приема'
      parameters:
        -
          name: per-page
          in: query
          description: 'Кол-во записей на странице'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
  '/api/reception-point/{id}':
    get:
      tags:
        - 'Пункты приема'
      summary: 'Найти пункт приема по id'
      parameters:
        -
          name: id
          in: path
          description: 'ID пунка приема'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
    put:
      tags:
        - 'Пункты приема'
      summary: 'Обновить пункт приема'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                title:
                  description: Название
                  type: string
                id_technologist:
                  description: 'Id технолога'
                  type: integer
                id_technician:
                  description: 'Id техника'
                  type: integer
                stream_url:
                  description: 'Ссылка на транслацию'
                  type: string
                status:
                  description: Статус
                  type: integer
                queue_vagons:
                  description: 'Очередь вагонов'
                  type: json
                id_order:
                  description: 'Id заказа'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
    delete:
      tags:
        - 'Пункты приема'
      summary: 'Удалить пункт приема по id'
      operationId: 7d29e4cf45409eb4386f1aa0b567ddb9
      parameters:
        -
          name: id
          in: path
          description: 'ID пункта'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
  /api/reception-point/:
    post:
      tags:
        - 'Пункты приема'
      summary: 'Добавить пункт приема'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                title:
                  description: Название
                  type: string
                id_technologist:
                  description: 'Id технолога'
                  type: integer
                id_technician:
                  description: 'Id техника'
                  type: integer
                stream_url:
                  description: 'Ссылка на транслацию'
                  type: string
                status:
                  description: Статус
                  type: integer
                queue_vagons:
                  description: 'Очередь вагонов'
                  type: json
                id_order:
                  description: 'Id заказа'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
...